pyproject.toml
[project]
name = "swiss-ephemeris-api"
version = "0.1.0"
description = "REST API for Swiss Ephemeris astronomical calculations"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.12"
keywords = ["astrology", "ephemeris", "swiss-ephemeris", "astronomy"]
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dateutil>=2.9.0",
    "pysweph>=2.10.3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
